+++ index.js
@@ -2,53 +2,10 @@ import React from 'react'
 import {usePersistReducer} from 'use-persist'
 import ReactDOM from 'react-dom'
 import styled from 'styled-components'
-import {sample, some} from 'lodash'
-import hiraganaMap from './src/lib/hiragana'
-import katakanaMap from './src/lib/katakana'
+import {some} from 'lodash'
 import {useHotkeys} from 'react-hotkeys-hook'
-import wordData from './words.json'
-import {fromKana, toHiragana, toKatakana, containsHiragana} from 'hepburn'
-import Speech from 'speak-tts'
 import Layout from './src/Layout'
-import {Howl} from 'howler'
-
-const correctSound = new Howl({
-  src: ['blip.mp3']
-})
-
-const speech = new Speech()
-speech.init({lang: 'ja'}).catch(console.error)
-
-const hiragana = Object.entries(hiraganaMap).map(([kana, romaji]) => ({
-  kana,
-  romaji
-}))
-const katakana = Object.entries(katakanaMap).map(([kana, romaji]) => ({
-  kana,
-  romaji
-}))
-
-const words = wordData.map((data) => ({
-  meaning: data['Vocab-meaning'],
-  romaji: fromKana(data['Vocab-kana'])
-    .replace('ē', 'ee')
-    .replace('ō', 'oo')
-    .replace("'", ''),
-  kana: data['Vocab-kana'],
-  expression: data['Vocab-expression']
-}))
-
-const initialState = {
-  input: '',
-  isWrong: false,
-  isRevealing: false,
-  correct: 0,
-  settings: {
-    hiragana: true,
-    katakana: true,
-    words: false
-  }
-}
+import {reducer, init, initialState} from './src/state'
 
 function App() {
   const [state, dispatch] = usePersistReducer(
@@ -117,104 +74,6 @@ function App() {
   function toggleReveal() {
     dispatch({type: 'toggleReveal'})
   }
-}
-
-function reducer(state, action) {
-  const {input, current, settings, isRevealing, correct} = state
-
-  console.log(action.type)
-
-  if (action.type === 'newPrompt') {
-    return {
-      ...state,
-      settings: action.settings || settings
-    }
-  }
-
-  if (
-    action.type === 'toggleReveal' ||
-    (action.type === 'onPressEnter' && isRevealing)
-  ) {
-    return {
-      ...state,
-      input: '',
-      isRevealing: !isRevealing
-    }
-  }
-
-  if (action.type === 'input') {
-    if (isCorrect(action.value, current)) return onCorrect()
-    if (action.value.includes('?')) return {...state, isRevealing: true}
-
-    return {
-      ...state,
-      input: action.value
-    }
-  }
-
-  if (action.type === 'onPressEnter' && !isRevealing) {
-    if (input.length === 0)
-      return {
-        ...state,
-        isWrong: false
-      }
-    if (isCorrect(input, current)) return onCorrect()
-    return onFailure()
-  }
-
-  if (action.type === 'setSettings') {
-    return {
-      ...state,
-      current: getPrompt(action.settings) || current,
-      settings: action.settings
-    }
-  }
-
-  return state
-
-  function onCorrect() {
-    correctSound.play()
-    speech.speak({text: current.kana})
-    return {
-      ...state,
-      correct: correct + 1,
-      current: getPrompt(settings),
-      isWrong: false,
-      input: ''
-    }
-  }
-
-  function onFailure() {
-    const romaji = input.trim().toUpperCase()
-    const usedKana = containsHiragana(current.kana)
-      ? toHiragana(romaji)
-      : toKatakana(romaji)
-    return {
-      ...state,
-      isWrong: `${romaji.toLowerCase()} (${usedKana || ''})`,
-      setInput: ''
-    }
-  }
-}
-
-function isCorrect(input, current) {
-  return input.toLowerCase().trim() === current.romaji.toLowerCase()
-}
-
-function getPrompt(settings) {
-  const category = sample(
-    [
-      settings.hiragana && hiragana,
-      settings.katakana && katakana,
-      settings.words && words
-    ].filter((list) => list)
-  )
-
-  return sample(category)
-}
-
-function init(state) {
-  return {current: getPrompt(state.settings), input: '', ...state}
 }
 
 const Prompt = styled.div`
